+++ src/hotspot/share/code/dependencies.cpp
@@ -780,7 +780,7 @@ void Dependencies::write_dependency_to(xmlStream* xtty,
       char xn[12];
       int printed_len = os::snprintf(xn, sizeof(xn), "x%d", j);
       assert(printed_len > 0, "error occurs at os::snprintf");
-      assert(printed_len < sizeof(xn), "xn buf overflow");
+      assert((size_t)printed_len < sizeof(xn), "xn buf overflow");
       if (arg.is_oop()) {
         xtty->object(xn, Handle(thread, arg.oop_value()));
       } else {
